+++ Source/Renderer/BlockRenderer.js
@@ -38,12 +38,15 @@ EJS.Renderer.BlockRenderer = new Class({
 	 * @return {Element} The renderer output
 	 */
 	render : function(language, specialLines, localOptions){
-		// create new outer container element - use ol tag if lineNumbers are enabled. element attribute settings are priorized
+        // elememt shortcut
+        var _el = EJS.Dom.Element;
+
+        // create new outer container element - use ol tag if lineNumbers are enabled. element attribute settings are priorized
 		var container = null;
 		if (localOptions.lineNumbers != null){
-			container = new EJS.Dom.Element((localOptions.lineNumbers.toLowerCase() === 'true') ? 'ol' : 'ul');
+			container = new _el((localOptions.lineNumbers.toLowerCase() === 'true') ? 'ol' : 'ul');
 		}else{
-			container = new EJS.Dom.Element(this.options.showLinenumbers ? 'ol' : 'ul');
+			container = new _el(this.options.showLinenumbers ? 'ol' : 'ul');
 		}
 		
 		// add "start" attribute ?
@@ -53,76 +56,65 @@ EJS.Renderer.BlockRenderer = new Class({
 		
 		// line number count
 		var lineCounter = 1;
-		
-		// current line element
-		var currentLine = new EJS.Dom.Element('li', {
-			'class': (specialLines.isSpecialLine(lineCounter) ? 'specialline' : '')
-		});
+
+        var tokens = language.getTokens();
+        var odd = ' ' + this.options.oddClassname || '';
+        var even = ' ' + this.options.evenClassname || '';
+
+        // current line element
+        var currentLine = new _el('li', {
+            'class': (specialLines.isSpecialLine(lineCounter) ? 'specialline' : '') + odd
+        });
 
         // output filter
-        var _F = (function(t){
-            return this.textFilter.filterOutput(t);
+        var addFragment = (function(className, text){
+            currentLine.grab(new _el('span', {
+                'class': className,
+                'text': this.textFilter.filterOutput(text)
+            }));
         }.bind(this));
-		
-		// generate output based on ordered list of tokens
-		language.getTokens().each(function(token, index){
-			// get classname
-			var className = token.type ? (language.aliases[token.type] || token.type) : '';
-			
-			// split the token into lines
-			var lines = token.text.split('\n');
-			
-			// linebreaks found ?
-			if (lines.length > 1){
-				// just add the first line
-				currentLine.grab(new EJS.Dom.Element('span', {
-					'class': className,
-					'text': _F(lines.shift())
-				}));
-				
-				// generate element for each line
-				lines.each(function(line, lineNumber){
-					// grab old line into output container
-					container.grab(currentLine);
-					
-					// new line
-					lineCounter++;
-					
-					// create new line, add special line classes
-					currentLine = new EJS.Dom.Element('li', {
-						'class': (specialLines.isSpecialLine(lineCounter) ? 'specialline' : '')
-					});
-										
-					// create new token-element
-					currentLine.grab(new EJS.Dom.Element('span', {
-						'class': className,
-						'text': _F(line)
-					}));
-				});				
-			}else{
-				// just add the token
-				currentLine.grab(new EJS.Dom.Element('span', {
-					'class': className,
-					'text': _F(token.text)
-				}));	
-			}			
-		});
-		
+
+        // generate output based on ordered list of tokens
+        Array.each(tokens, function(token){
+            // get classname
+            var className = token.type ? (language.aliases[token.type] || token.type) : '';
+
+            // split the token into lines
+            var lines = token.text.split('\n');
+
+            // linebreaks found ?
+            if (lines.length > 1){
+                // just add the first line
+                addFragment(className, lines.shift());
+
+                // generate element for each line
+                Array.each(lines, function(line, lineNumber){
+                    // grab old line into output container
+                    container.grab(currentLine);
+
+                    // new line
+                    lineCounter++;
+
+                    // create new line, add special line classes; add odd/even classes
+                    currentLine = new _el('li', {
+                        'class': (specialLines.isSpecialLine(lineCounter) ? 'specialline' : '') + (lineCounter%2==0 ? even: odd)
+                    });
+
+                    // create new token-element
+                    addFragment(className, line);
+                });
+            }else{
+                addFragment(className, token.text);
+            }
+        });
+
 		// grab last line into container
 		container.grab(currentLine);
-
-		// add odd/even classes
-		if (this.options.evenClassname){
-			container.getElements('li:even').addClass(this.options.evenClassname);
-		}
-		if (this.options.oddClassname){
-			container.getElements('li:odd').addClass(this.options.oddClassname);
-		}
 
 		// highlight lines ?
 		if (this.options.hover && this.options.hover != "NULL"){
 			// add hover enable class
-			container.getChildren().addClass(this.options.hover);
+			container.addClass(this.options.hover);
 		}
 
 		return container;