+++ Util/ExampleGenerator.php
@@ -11,9 +11,11 @@ function cdnbase($file){
 // Get output dir
 $outputDir = (count($argv)==2 ? $argv[1] : 'Output/');
 
+require('global.php');
+
 // === Basic Examples =================================================
 $examples = array(
-    'Example1', 'Example2-jsinit', 'Example3-advanced', 'LanguageDevelopment', 'ThemeDevelopment'
+    'Example1', 'Example2-jsinit', 'Example3-advanced'
 );
 foreach ($examples as $example){
     file_put_contents('Examples/'.$example.'.html', captureTemplate('Resources/ExampleData/'.$example.'.phtml', array(
@@ -22,10 +24,7 @@ foreach ($examples as $example){
 }
 
 // === Language Examples =================================================
-$languages = array(
-	'C', 'Cpp', 'CSharp', 'CSS', 'Diff', 'HTML', 'Java', 'Javascript', 'JSON', 'MarkDown', 'NSIS', 'PHP', 'Python', 'Ruby', 'SQL', 'Unit', 'XML', 'RAW', 'NoHighlight', 'Ini', 'AVRASM'
-);
-foreach ($languages as $lang){
+foreach ($languageExamples as $lang){
 	$langContent = file_get_contents('Resources/TestcaseData/'.strtolower($lang).'.html');
 	renderTemplate($outputDir.strtolower($lang).'.html', array(
 		'pageContent' => $langContent,
@@ -39,19 +38,3 @@ foreach ($languages as $lang){
 function renderTemplate($destination, $vars = array()){
     file_put_contents($destination, captureTemplate('Resources/Templates/ExampleTemplate.phtml', $vars));
 }
-
-function captureTemplate($file, $vars = array()){
-    // exapand vars to local variables
-    extract($vars);
-
-    // start capturing
-    ob_start();
-
-    // load local template file
-    require($file);
-
-    // store captured content
-    $_generatedContent = ob_get_clean();
-    return $_generatedContent;
-}
-?>