+++ Util/PageGenerator.php
@@ -16,13 +16,7 @@ $outputDir = (count($argv)>=2 ? $argv[1] : 'Output/');
 // Get Build Version
 define('ENLIGHTERJS_VERSION', (count($argv)>=3 ? $argv[2] : 'unknown'));
 
-$themes = array(
-    'Enlighter', 'Classic', 'Eclipse', 'Beyond', 'Git', 'Mocha', 'MooTools', 'Panic', 'Tutti', 'Twilight'
-);
-
-$languages = array(
-    'C', 'Cpp', 'CSharp', 'CSS', 'Diff', 'HTML', 'Java', 'Javascript', 'JSON', 'MarkDown', 'NSIS', 'PHP', 'Python', 'Ruby', 'SQL', 'Unit', 'XML', 'RAW', 'NoHighlight', 'AVRASM', 'Ini'
-);
+require('global.php');
 
 // === README ============================================================
 // create Readme.html
@@ -55,7 +49,10 @@ $frontpageContent = captureTemplate('Resources/Templates/Frontpage.phtml');
 file_put_contents($outputDir.'index.html', $frontpageContent);
 
 // === Builder ========================================================
-$builderContent = captureTemplate('Resources/Templates/Builder.phtml') ;
+$builderContent = captureTemplate('Resources/Templates/Builder.phtml', array(
+    'languages' => $languageDescriptions,
+    'themes' => $themes
+));
 renderTemplate($outputDir.'Builder.html', array(
 'pageContent' => $builderContent,
 'pageTitle' => 'Builder <small>Generate your custom EnlighterJS package</small>'
@@ -63,7 +60,7 @@ renderTemplate($outputDir.'Builder.html', array(
 
 
 // === Language Examples =================================================
-foreach ($languages as $lang){
+foreach ($languageList as $lang){
     $langContent = captureTemplate('Resources/Templates/ThemeSelector.phtml', array(
         'themes' => $themes
     )) ;
@@ -90,61 +87,45 @@ renderTemplate($outputDir.'Themes.html', array(
 ));
 
 
-	function extractContent($filename, $start=0, $stop=99999){
-		$content = file_get_contents($filename);
-		$lines = explode("\n", $content);
-		$buffer = '';
+function extractContent($filename, $start=0, $stop=99999){
+    $content = file_get_contents($filename);
+    $lines = explode("\n", $content);
+    $buffer = '';
 
-		for ($i=$start;$i<min($stop, count($lines)-1);$i++){
-			$buffer .= $lines[$i]. "\n";
-		}
-		
-		return $buffer;
-	}
+    for ($i=$start;$i<min($stop, count($lines)-1);$i++){
+        $buffer .= $lines[$i]. "\n";
+    }
 
-	/**
-	 * Renders the template file and return HTML
-	 * @param Array $vars
-	 */
-	function renderTemplate($destination, $vars = array()){
-		file_put_contents($destination, captureTemplate('Resources/Templates/WebsiteTemplate.phtml', $vars));
-	}
-	
-	function captureTemplate($file, $vars = array()){
-		// exapand vars to local variables
-		extract($vars);
-		
-		// start capturing
-		ob_start();
-		
-		// load local template file
-		require($file);
-		
-		// store captured content
-		$_generatedContent = ob_get_clean();
-		return $_generatedContent;
-	}
+    return $buffer;
+}
 
-	/**
-	 * Render a Markdown Document using LightUp with Promethium CloudAPI
-	 * @param unknown $content
-	 */
-	function renderMarkdownDocument($content) {
-		$postdata = http_build_query(array(
-				'mddata' => $content,
-				'highlightingMode' => 'enlighterjs',
-				'addAnchors' => 'true'
-		));
-		$opts = array(
-				'http' => array (
-						'method' => 'POST',
-						'header' => 'Content-type: application/x-www-form-urlencoded',
-						'content' => $postdata
-				)
-		);
-		$htmlContent = file_get_contents('http://promethium.andidittrich.de/lightup/', false, stream_context_create($opts));
-		
-		// remove first heading1
-		return preg_replace('/<h1>.*<\/h1>/', '', $htmlContent, 1);
-	}
-?>
+/**
+ * Renders the template file and return HTML
+ * @param Array $vars
+ */
+function renderTemplate($destination, $vars = array()){
+    file_put_contents($destination, captureTemplate('Resources/Templates/WebsiteTemplate.phtml', $vars));
+}
+
+/**
+ * Render a Markdown Document using LightUp with Promethium CloudAPI
+ * @param unknown $content
+ */
+function renderMarkdownDocument($content) {
+    $postdata = http_build_query(array(
+            'mddata' => $content,
+            'highlightingMode' => 'enlighterjs',
+            'addAnchors' => 'true'
+    ));
+    $opts = array(
+            'http' => array (
+                    'method' => 'POST',
+                    'header' => 'Content-type: application/x-www-form-urlencoded',
+                    'content' => $postdata
+            )
+    );
+    $htmlContent = file_get_contents('http://promethium.andidittrich.de/lightup/', false, stream_context_create($opts));
+
+    // remove first heading1
+    return preg_replace('/<h1>.*<\/h1>/', '', $htmlContent, 1);
+}